# sdk/go/api/openapi.yaml
openapi: 3.1.0
info:
  description: "C2PA-compliant content signing and verification infrastructure for\
    \ publishers, legal/finance firms, AI labs, and enterprises"
  title: Encypher Enterprise API
  version: 1.0.0-preview
servers:
- description: Production
  url: https://api.encypherai.com
- description: Local development
  url: http://localhost:8007
paths:
...
  /api/v1/webhooks:
    get:
      description: List all webhooks for the organization.
      operationId: list_webhooks_api_v1_webhooks_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List Webhooks
      tags:
      - Webhooks
      - Webhooks
    post:
      description: |-
        Register a new webhook.

        The webhook URL must be HTTPS. You can optionally provide a shared secret
        for HMAC signature verification of webhook payloads.
      operationId: create_webhook_api_v1_webhooks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookCreateRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookCreateResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Webhook
      tags:
      - Webhooks
      - Webhooks
  /api/v1/webhooks/{webhook_id}:
    delete:
      description: Delete a webhook.
      operationId: delete_webhook_api_v1_webhooks__webhook_id__delete
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookDeleteResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete Webhook
      tags:
      - Webhooks
      - Webhooks
    get:
      description: Get details of a specific webhook.
      operationId: get_webhook_api_v1_webhooks__webhook_id__get
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Webhook
      tags:
      - Webhooks
      - Webhooks
    patch:
      description: "Update a webhook's URL, events, or active status."
      operationId: update_webhook_api_v1_webhooks__webhook_id__patch
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookUpdateResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update Webhook
      tags:
      - Webhooks
      - Webhooks
  /api/v1/webhooks/{webhook_id}/test:
    post:
      description: |-
        Send a test event to the webhook.

        This sends a test payload to verify the webhook is configured correctly.
      operationId: test_webhook_api_v1_webhooks__webhook_id__test_post
      parameters:
      - explode: false
        in: path
        name: webhook_id
        required: true
        schema:
          title: Webhook Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTestResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Test Webhook
      tags:
      - Webhooks
      - Webhooks
...
components:
  schemas:
...
    WebhookTestResponse:
      description: Response after testing a webhook.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: WebhookTestResponse
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    WebhookCreateResponse:
      description: Response after creating a webhook.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: WebhookCreateResponse
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WebhookUpdateRequest:
      description: Request to update a webhook.
      example:
        is_active: true
        url: url
        events:
        - events
        - events
      properties:
        url:
          nullable: true
          type: string
        events:
          items:
            type: string
          nullable: true
          type: array
        is_active:
          nullable: true
          type: boolean
      title: WebhookUpdateRequest
...
    WebhookCreateRequest:
      description: Request to create a webhook.
      example:
        secret: secret
        url: url
        events:
        - events
        - events
      properties:
        url:
          description: Webhook URL (must be HTTPS)
          maxLength: 2048
          title: Url
          type: string
        events:
          description: Events to subscribe to
          items:
            type: string
          minItems: 1
          type: array
        secret:
          maxLength: 256
          minLength: 16
          nullable: true
          type: string
      required:
      - events
      - url
      title: WebhookCreateRequest
...
    WebhookDeleteResponse:
      description: Response after deleting a webhook.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: WebhookDeleteResponse
...
    WebhookUpdateResponse:
      description: Response after updating a webhook.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: WebhookUpdateResponse
...
    WebhookListResponse:
      description: Response for webhook listing.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: WebhookListResponse
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http